Move About page to app router
