--- app/about/(misc)/index.ts
@@ -1,5 +1,0 @@
-export { default as Year2019 } from "./2019.jpg";
-export { default as Year2020 } from "./2020.jpg";
-export { default as Year2021 } from "./2021.jpg";
-export { default as Year2022 } from "./2022.jpg";
-export { default as ImpactReport } from "./impact.jpg";
