--- app/about/(team)/index.ts
@@ -1,20 +1,0 @@
-export { default as Anezka } from "./anezka.jpg";
-export { default as Eva } from "./eva.jpg";
-export { default as Gabi } from "./gabi.jpg";
-export { default as Jindra } from "./jindra.jpg";
-export { default as Lukas } from "./lukas.jpg";
-export { default as Martina } from "./martina.jpg";
-export { default as Romana } from "./romana.jpg";
-export { default as Terezia } from "./terezia.jpg";
-export { default as Terka } from "./terka.jpg";
-export { default as Yen } from "./yen.jpg";
-export { default as Zoul } from "./zoul.jpg";
-export { default as Lujza } from "./lujza.jpg";
-export { default as Tereza } from "./tereza.jpg";
-export { default as Veronika } from "./veronika.jpg";
-
-export { default as Michal } from "./michal.jpg";
-export { default as Jakub } from "./jakub.jpg";
-export { default as Robert } from "./robert.jpg";
-export { default as Ondrej } from "./ondrej.jpg";
-export { default as Dita } from "./dita.jpg";
