--- app/about/(volunteers)/index.ts
@@ -1,9 +1,0 @@
-export { default as Tereza } from "./tereza.png";
-export { default as Tomas } from "./tomas.png";
-export { default as TomasN } from "./tomasn.png";
-export { default as Radka } from "./radka.png";
-export { default as Jan } from "./jan.png";
-export { default as Katerina } from "./katerina.png";
-export { default as Lucia } from "./lucia.png";
-export { default as Ondrej } from "./ondrej.png";
-export { default as RadkaH } from "./radkah.png";
