+++ app/about/misc/index.ts
@@ -1,0 +1,15 @@
+import Year2019 from "./2019.jpg";
+import Year2020 from "./2020.jpg";
+import Year2021 from "./2021.jpg";
+import Year2022 from "./2022.jpg";
+import ImpactReport from "./impact.jpg";
+
+const Images = {
+  Year2019,
+  Year2020,
+  Year2021,
+  Year2022,
+  ImpactReport,
+};
+
+export default Images;
