+++ app/about/team/index.ts
@@ -1,0 +1,43 @@
+import Anezka from "./anezka.jpg";
+import Eva from "./eva.jpg";
+import Gabi from "./gabi.jpg";
+import Jindra from "./jindra.jpg";
+import Lukas from "./lukas.jpg";
+import Martina from "./martina.jpg";
+import Romana from "./romana.jpg";
+import Terezia from "./terezia.jpg";
+import Terka from "./terka.jpg";
+import Yen from "./yen.jpg";
+import Zoul from "./zoul.jpg";
+import Lujza from "./lujza.jpg";
+import Tereza from "./tereza.jpg";
+import Veronika from "./veronika.jpg";
+import Michal from "./michal.jpg";
+import Jakub from "./jakub.jpg";
+import Robert from "./robert.jpg";
+import Ondrej from "./ondrej.jpg";
+import Dita from "./dita.jpg";
+
+const Images = {
+  Anezka,
+  Eva,
+  Gabi,
+  Jindra,
+  Lukas,
+  Martina,
+  Romana,
+  Terezia,
+  Terka,
+  Yen,
+  Zoul,
+  Lujza,
+  Tereza,
+  Veronika,
+  Michal,
+  Jakub,
+  Robert,
+  Ondrej,
+  Dita,
+};
+
+export default Images;
