+++ app/about/team/jakub.jpg
