+++ app/about/team/jindra.jpg
