+++ app/about/team/michal.jpg
