+++ app/about/volunteers/index.ts
@@ -1,0 +1,23 @@
+import Tereza from "./tereza.png";
+import Tomas from "./tomas.png";
+import TomasN from "./tomasn.png";
+import Radka from "./radka.png";
+import Jan from "./jan.png";
+import Katerina from "./katerina.png";
+import Lucia from "./lucia.png";
+import Ondrej from "./ondrej.png";
+import RadkaH from "./radkah.png";
+
+const Images = {
+  Tereza,
+  Tomas,
+  TomasN,
+  Radka,
+  Jan,
+  Katerina,
+  Lucia,
+  Ondrej,
+  RadkaH,
+};
+
+export default Images;
